+++ provider/pkg/kinds/util.go
@@ -4,11 +4,12 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
 )
 
-// IsPatchURN returns true if the URN is for a Patch resource.
-func IsPatchURN(urn resource.URN) bool {
-	urnS := urn.Type().String()
+// IsPatchResource returns true if it is a Patch resource and also ensures
+// that we don't return true for CustomResources that end with "Patch".
+func IsPatchResource(urn resource.URN, kind string) bool {
+	resourceName := urn.Type().Name().String()
 
-	return PatchQualifiedTypes.Has(urnS)
+	return kind+"Patch" == resourceName
 }
 
 // IsListURN returns true if the URN is for a List resource.